feat: move user logic
