+++ src/app/error.tsx
@@ -9,7 +9,7 @@ export default function Error({
   reset: () => void
 }) {
   useEffect(() => {
-    // Log the error to an error reporting service
+    // TODO: Log the error to an error reporting service
     console.error(error)
   }, [error])
