+++ src/app/styles/Home.module.css
@@ -1,59 +1,29 @@
 .container {
   display: flex;
   flex-direction: column;
-  height: 100vh;
+  min-height: 100vh;
 }
 
-.main {
-  flex: 1;
-  padding: 0 1rem;
-  display: flex;
-  flex-direction: column;
+.nextStepsCard {
+  margin: 0 auto;
+  width: 600px;
 }
 
 .wrapper {
   display: flex;
   flex-direction: column;
-  gap: 2rem;
-}
-
-.frost {
-  background: rgba(255, 255, 255, 0.2);
-  border-radius: 16px;
-  box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
-  backdrop-filter: blur(5px);
-  -webkit-backdrop-filter: blur(5px);
-  border: 1px solid rgba(255, 255, 255, 0.31);
 }
 
 .header {
   display: flex;
-  flex-direction: column;
-  padding: 4rem 0 2rem 0;
+  margin: auto;
   text-align: center;
-  gap: 2rem;
-}
-
-.header h1 {
-  max-width: 620px;
-  align-self: center;
-}
-
-.body {
-  composes: frost;
-  border-radius: var(--border-radius);
-  display: flex;
-  padding: 1rem;
-  width: 100%;
-  flex-direction: row;
-  align-self: center;
-  max-width: min(800px, var(--screen-width));
+  gap: 4rem;
 }
 
 .instructions {
   display: inline-flex;
   flex-direction: column;
-  padding: 1rem;
   width: 100%;
 }
 
@@ -76,19 +46,27 @@
 }
 
 .footer {
-  display: flex;
-  padding: 2rem 0;
-  justify-content: center;
   align-items: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  padding: 2rem 0;
 }
 
-.footer span {
-  margin-right: 0.2rem;
+.footer a,
+.footer p {
+  margin: auto 0.4rem;
+}
+
+.footer-links {
+  display: flex;
+  gap: 1rem;
 }
 
 .code {
   display: flex;
   align-items: center;
+  margin: 0.5rem;
   gap: 0.5rem;
 }
 
@@ -104,33 +82,6 @@
   align-items: center;
   justify-content: center;
   color: rgb(57, 33, 97);
-}
-
-.code pre {
-  background: rgba(124, 58, 237, 12%);
-  border-radius: 4px;
-  padding: 0.7em 1.4em;
-  text-align: center;
-}
-
-.code code {
-  font-size: 0.86em;
-  font-weight: bold;
-  color: rgb(124, 58, 237);
-  font-family: Menlo, Monaco, Lucida Console, Liberation Mono, DejaVu Sans Mono,
-    Bitstream Vera Sans Mono, Courier New, monospace;
-}
-
-.toastContainer {
-  border: 1px solid #edff;
-  padding: 0 1rem;
-  background: #eeff;
-  color: #62af;
-  text-align: center;
-}
-
-.toastContainer strong {
-  color: rgb(124, 58, 237);
 }
 
 .textLink {
@@ -156,42 +107,6 @@
 .textLink:focus,
 .textLink:active {
   background-size: 0 1px, 100% 1px;
-}
-
-.arrowIcon {
-  box-sizing: border-box;
-  display: block;
-  width: 8px;
-  height: 8px;
-  border-top: 2px solid;
-  transform: scale(var(--ggs, 1));
-  border-right: 2px solid;
-  position: absolute;
-  right: 6px;
-  top: 6px;
-  color: #b1a5c4;
-}
-
-.arrowIcon::after {
-  content: "";
-  display: block;
-  box-sizing: border-box;
-  position: absolute;
-  width: 8px;
-  height: 2px;
-  background: currentColor;
-  transform: rotate(-45deg);
-  top: 2px;
-  right: -1px;
-}
-
-.buttonContainer {
-  display: flex;
-  flex-direction: row;
-  gap: 1rem;
-  justify-content: center;
-  align-items: center;
-  flex: 1;
 }
 
 .button {
@@ -223,90 +138,23 @@
   box-shadow: rgb(80 63 205 / 50%) 0px 1px 40px;
 }
 
-.loginButton {
-  composes: button;
-  background: rgb(248 250 252);
-  border: 1px solid rgb(231, 216, 246);
-  color: rgb(30 41 59);
-  text-shadow: none;
-}
-
-.loginButton:hover {
-  color: rgb(30 41 59);
-  text-shadow: none;
-}
-
-.card:hover .arrowIcon {
-  color: #7450ec;
-}
-
-.linkGrid {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-  gap: 1rem;
-}
-
-.card {
-  composes: frost;
-  padding: 1rem 0rem;
-  text-align: center;
-  color: inherit;
-  text-decoration: none;
-  border-radius: 10px;
-  border-bottom-left-radius: 0px;
-  transition: color 0.15s ease, border-color 0.15s ease;
-  max-width: 200px;
-  min-width: 200px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-}
-
-.card:hover,
-.card:focus,
-.card:active {
-  color: #7450ec;
-  border-color: #7450ec;
-}
-
-.card h2 {
-  margin: 0 0 1rem 0;
-  font-size: 1.5rem;
-}
-
-.card p {
-  margin: 0;
-  font-size: 1.25rem;
-  line-height: 1.5;
-}
-
-.logo {
-  flex: 1;
-  padding: 1rem 2rem;
-}
-
-.logo svg {
-  height: 100%;
-  width: 200px;
-  fill: #7450ec;
+.logo img {
+  border-radius: 16px;
+  box-shadow: rgba(117, 81, 236, 0.8) 3px 6px 18px;
+  margin: auto;
 }
 
 /* MOBILE */
 @media (max-width: 800px) {
-  .linkGrid {
-    width: 100%;
+  .footer {
+    flex-direction: column;
   }
 
-  .card {
-    max-width: 100%;
+  .header {
+    flex-direction: column;
   }
 
-  .body {
-    flex-wrap: wrap;
-  }
-
-  .buttonContainer {
-    flex-wrap: wrap;
+  .nextStepsCard {
+    width: 300px;
   }
 }
