+++ src/app/styles/globals.css
@@ -7,8 +7,6 @@ html,
 body {
   padding: 0;
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
-    Ubuntu, Cantarell, "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
 }
 
 a {
@@ -22,4 +20,8 @@ i {
 
 * {
   box-sizing: border-box;
+}
+
+.font-special {
+  font-family: "Courier New", Courier, monospace;
 }
