+++ src/app/users/queries/getCurrentUser.ts
@@ -5,7 +5,16 @@ export default async function getCurrentUser(_: null, ctx: Ctx) {
   if (!ctx.session.userId) return null
   const user = await db.user.findFirst({
     where: { id: ctx.session.userId },
-    select: { id: true, name: true, email: true, role: true },
+    select: {
+      id: true,
+      nameFirst: true,
+      nameLast: true,
+      email: true,
+      emailBackup: true,
+      emailStripe: true,
+      role: true,
+      subscriptions: true,
+    },
   })
 
   return user
