--- src/styles/globals.css
@@ -1,27 +1,0 @@
-:root {
-  --border-radius: 0.75rem;
-  --screen-width: 90vmin;
-}
-
-html,
-body {
-  padding: 0;
-  margin: 0;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-i {
-  font-size: 0.8rem;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-.font-special {
-  font-family: "Courier New", Courier, monospace;
-}
