--- src/users/hooks/useCurrentUser.ts
@@ -1,7 +1,0 @@
-import { useQuery } from "@blitzjs/rpc"
-import getCurrentUser from "src/users/queries/getCurrentUser"
-
-export const useCurrentUser = () => {
-  const [user] = useQuery(getCurrentUser, null)
-  return user
-}
