--- src/users/queries/getCurrentUser.ts
@@ -1,22 +1,0 @@
-import { Ctx } from "blitz"
-import db from "db"
-
-export default async function getCurrentUser(_ = null, { session }: Ctx) {
-  if (!session.userId) return null
-
-  const user = await db.user.findFirst({
-    where: { id: session.userId },
-    select: {
-      id: true,
-      nameFirst: true,
-      nameLast: true,
-      email: true,
-      emailBackup: true,
-      emailStripe: true,
-      role: true,
-      subscriptions: true,
-    },
-  })
-
-  return user
-}
